+++ app/Http/Requests/Admin/ServerSave.php
@@ -116,6 +116,7 @@ class ServerSave extends FormRequest
             'spectific_key' => 'nullable|string',
             'code' => 'nullable|string',
             'show' => '',
+            'auto_online' => 'nullable|boolean',
             'name' => 'required|string',
             'group_ids' => 'nullable|array',
             'route_ids' => 'nullable|array',